Handles first boot case
Measurements are still subject to noise because we still do things after
registration but before the first interactive boot:

1. touch the filesystem and
2. create a user after registration

It seems that I found a pattern to replicate in CI jobs that require
unregistering WSL instnaces:

1. shutdown the VM
2. wait a little bit
3. unregister

With that pattern I zeroed out the cases of hangs due wslservice.exe
crashes.

I reduced the number of samples per relase x case and removed 20.04
from the sampling process to save time, as Focal cannot run cloud-init
as it is now.